Move CustomConsoleHandler to gui.py
